requirements.txt
backcall==0.1.0
certifi==2019.9.11
chardet==3.0.4
decorator==4.4.1
Django==2.2.7
djangorestframework==3.10.3
idna==2.8
ipython==7.9.0
ipython-genutils==0.2.0
jedi==0.15.1
oauthlib==3.1.0
parso==0.5.1
pexpect==4.7.0
pickleshare==0.7.5
prompt-toolkit==2.0.10
ptyprocess==0.6.0
Pygments==2.5.1
pytz==2019.3
requests==2.22.0
requests-oauthlib==1.3.0
six==1.13.0
sqlparse==0.3.0
traitlets==4.3.3
urllib3==1.25.7
wcwidth==0.1.7
elasticsearch>=6.0.0,<7.0.0